conversational-ai-pipeline/requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
pyaudio>=0.2.11
opencv-python>=4.8.0
mss>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
webrtcvad>=2.0.10
transformers>=4.30.0
datasets>=2.12.0
librosa>=0.10.0
